# Conflicts:
#	app/src/main/java/lab/infoworks/starter/ui/activities/app/AppViewModel.java
#	app/src/main/java/lab/infoworks/starter/ui/activities/riderList/RiderList.java
#	libshared/src/main/java/lab/infoworks/libshared/domain/repository/definition/RiderRepository.java
#	libshared/src/main/java/lab/infoworks/libshared/domain/repository/impl/RiderRepositoryImpl.java

+++ libshared/src/main/java/lab/infoworks/libshared/domain/datasource/RiderDataSource.java
@@ -2,6 +2,7 @@ package lab.infoworks.libshared.domain.datasource;
 
 import android.content.Context;
 import android.os.Build;
+import android.util.Log;
 
 import androidx.annotation.RequiresApi;
 
@@ -16,11 +17,30 @@ import java.util.function.Consumer;
 import lab.infoworks.libshared.domain.db.AppDB;
 import lab.infoworks.libshared.domain.db.dao.RiderDAO;
 import lab.infoworks.libshared.domain.model.Rider;
+import lab.infoworks.libshared.domain.remote.BearerTokenInterceptor;
+import lab.infoworks.libshared.domain.remote.RemoteConfig;
+import lab.infoworks.libshared.domain.remote.api.RiderApiService;
+import okhttp3.Interceptor;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 public class RiderDataSource extends CMDataSource<Integer, Rider> implements DataStorage, AutoCloseable {
 
     private AppDB db;
     private ExecutorService executor;
+    private String baseUrl;
+    private RiderApiService apiService;
+
+    /**
+     * if (local_first == true)
+     * -> fetch: first get what was locally saved -> notify UI -> fire async-remote call -> update local store -> notify UI.
+     * -> insert/update/delete: first happen locally -> notify UI -> fire async-remote call.
+     * else
+     * -> fetch: first fetch from remote -> update local -> notify UI.
+     * -> insert/update/delete: first happen remote -> update in local -> notify UI.
+     */
+    private final boolean sync_state_local_first;
 
     private ExecutorService getExecutor() {
         if (executor == null){
@@ -29,17 +49,69 @@ public class RiderDataSource extends CMDataSource<Integer, Rider> implements Dat
         return executor;
     }
 
+    private RiderApiService getApiService() {
+        if (baseUrl == null || baseUrl.isEmpty()) return null;
+        if (apiService == null){
+            //Interceptor jwtToken = new BearerTokenInterceptor("jwt-token"); //Get From Some-where
+            apiService = RemoteConfig.getInstance(baseUrl, RiderApiService.class);
+        }
+        return apiService;
+    }
+
     public RiderDataSource(Context context){
+        this(context, "true", null);
+    }
+
+    public RiderDataSource(Context context, String localFirst, String baseUrl) {
         this.db = AppDB.getInstance(context);
-        retrieve();
+        this.sync_state_local_first = Boolean.parseBoolean(localFirst);
+        this.baseUrl = baseUrl;
+        if (sync_state_local_first){
+            retrieve();
+        }
     }
 
     @Override @RequiresApi(Build.VERSION_CODES.N)
     public void readAsync(int offset, int pageSize, Consumer<Rider[]> consumer) {
         if (consumer != null) {
-            //TODO:
-            List<Rider> items = readSyncAsList(offset, pageSize);
-            consumer.accept(items.toArray(new Rider[0]));
+            //
+            if (sync_state_local_first){
+                //Fetch Locally:
+                List<Rider> items = readSyncAsList(offset, pageSize);
+                //Notify UI:
+                consumer.accept(items.toArray(new Rider[0]));
+                //Fetch async from remote:
+                if (getApiService() != null){
+                    getApiService()
+                            .fetch(pageSize, offset)
+                            .enqueue(new Callback<List<Rider>>() {
+                                @Override
+                                public void onResponse(Call<List<Rider>> call, Response<List<Rider>> response) {
+                                    List<Rider> riders = response.body();
+                                    Log.d("RIDER-API", "Size: " + response.body().size());
+                                    int idx = items.size() - 1;
+                                    for (Rider rdr : riders) {
+                                        rdr.setId(++idx);
+                                        put(idx, rdr);
+                                    }
+                                    //Update Local-Store:
+                                    save(true);
+                                    //Notify UI Again
+                                    items.addAll(riders);
+                                    consumer.accept(items.toArray(new Rider[0]));
+                                }
+
+                                @Override
+                                public void onFailure(Call<List<Rider>> call, Throwable t) {
+                                    Log.d("RIDER-API-FAILED", t.getLocalizedMessage());
+                                }
+                            });
+                    //
+                }
+            }else {
+                //...
+                if (getApiService() != null){}
+            }
         }
     }
 
@@ -51,6 +123,8 @@ public class RiderDataSource extends CMDataSource<Integer, Rider> implements Dat
                 RiderDAO dao = db.riderDao();
                 dao.insert(new ArrayList<>(getInMemoryStorage().values()));
             });
+        }else {
+            //...
         }
     }
 
@@ -70,10 +144,16 @@ public class RiderDataSource extends CMDataSource<Integer, Rider> implements Dat
     @Override
     public boolean delete() {
         //TODO: Delete Data using Preferred Persistence Technology:
-        getExecutor().submit(() -> {
-            db.riderDao().deleteAll();
-        });
-        return true;
+        if (sync_state_local_first){
+            getExecutor().submit(() -> {
+                db.riderDao().deleteAll();
+            });
+            return true;
+        }else {
+            //...
+            if (getApiService() != null){}
+            return false;
+        }
     }
 
     @Override
